Update Austronesian drivers&makers
If a language hat a form in a concept, which was however lacking cognate coding,
that concept was coded as 000000
This is now fixed in Lexedata (not committed yet) and here's the new XML files
and code to generate them

Bad languages are those without coverage, i.e. without any cognate-coded forms
(and 831) – deriving this from the data would be better, probably by changing
coverage report to also say how many concepts are cognate-coded.